split flatbuffers.h into separate files
